up to 2147483647

+++ coap_callbacks.cpp
@@ -484,7 +484,7 @@ void callback_total(CoapPacket &packet, String url, IPAddress ip, int port) {
     // clear out packet fields and craft response
     clear_packet(packet);
 
-    int lng = currentStep;                   // up to stupid big number
+    long lng = currentStep;                   // up to stupid big number
     packet.payloadlen = 4;
     packet.payload[0] = (byte) ((lng & 0xFF000000) >> 24 );
     packet.payload[1] = (byte) ((lng & 0x00FF0000) >> 16 );
